+++ tests/test_delta_schunk.c
@@ -18,11 +18,11 @@ int main() {
   static int32_t data_dest[SIZE];
   size_t isize = SIZE * sizeof(int32_t);
   int dsize;
-  int64_t nbytes, cbytes;
+  size_t nbytes, cbytes;
   blosc2_cparams cparams = BLOSC_CPARAMS_DEFAULTS;
   blosc2_dparams dparams = BLOSC_DPARAMS_DEFAULTS;
   blosc2_schunk* schunk;
-  size_t nchunks;
+  int32_t nchunks;
 
   printf("Blosc version info: %s (%s)\n",
          BLOSC_VERSION_STRING, BLOSC_VERSION_DATE);
@@ -54,14 +54,14 @@ int main() {
   }
 
   /* Retrieve and decompress the chunks (0-based count) */
-  for (size_t nchunk = 0; nchunk < NCHUNKS; nchunk++) {
+  for (int nchunk = 0; nchunk < NCHUNKS; nchunk++) {
     dsize = blosc2_schunk_decompress_chunk(schunk, nchunk, (void *) data_dest, isize);
     if (dsize < 0) {
       return EXIT_FAILURE;
     }
     for (int i = 0; i < SIZE; i++) {
       if (data_dest[i] != i  * nchunk) {
-        fprintf(stderr, "First error in: %zu, %d, %d\n",
+        fprintf(stderr, "First error in: %u, %d, %d\n",
                 nchunk, i, data_dest[i]);
         return EXIT_FAILURE;
       }
